addHoldingsDialog

+++ holdings-tab.go
@@ -13,46 +13,47 @@ import (
 )
 
 func (app *Config) holdingsTab() *fyne.Container {
-	app.HoldingsTable = app.getHoldingsTable()
+	app.Holdings = app.getHoldingSlice()
+	app.HoldingsTable = app.buildHoldingsTable()
 
-	return container.NewBorder(
+	holdingsContainer := container.NewBorder(
 		nil,
 		nil,
 		nil,
 		nil,
 		container.NewAdaptiveGrid(1, app.HoldingsTable),
 	)
+
+	return holdingsContainer
+
 }
 
-func (app *Config) getHoldingsTable() *widget.Table {
-	data := app.getHoldingSlice()
-	app.Holdings = data
-
+func (app *Config) buildHoldingsTable() *widget.Table {
 	t := widget.NewTable(
 		func() (int, int) {
-			return len(data), len(data[0])
+			return len(app.Holdings), len(app.Holdings[0])
 		},
 		func() fyne.CanvasObject {
 			ctr := container.NewVBox(widget.NewLabel(""))
 			return ctr
 		},
 		func(i widget.TableCellID, o fyne.CanvasObject) {
-			if i.Col == (len(data[0])-1) && i.Row != 0 {
+			if i.Col == (len(app.Holdings[0])-1) && i.Row != 0 {
 				w := widget.NewButtonWithIcon("Delete", theme.DeleteIcon(),
 					func() {
-						app.showConfirmDeleteDialog(data, i)
+						app.showConfirmDeleteDialog(app.Holdings, i)
 					})
 				w.Importance = widget.HighImportance
 				o.(*fyne.Container).Objects = []fyne.CanvasObject{w}
 			} else {
 				o.(*fyne.Container).Objects = []fyne.CanvasObject{
-					widget.NewLabel(data[i.Row][i.Col].(string)),
+					widget.NewLabel(app.Holdings[i.Row][i.Col].(string)),
 				}
 			}
 		},
 	)
-	colWidths := []float32{50, 200, 200, 200, 110}
 
+	colWidths := []float32{50, 200, 200, 200, 110}
 	for i := range colWidths {
 		t.SetColumnWidth(i, colWidths[i])
 	}
@@ -98,11 +99,13 @@ func (app *Config) currentHoldings() ([]repository.Holdings, error) {
 
 func (app *Config) showConfirmDeleteDialog(data [][]interface{}, i widget.TableCellID) {
 	dialog.ShowConfirm("Delete", "", func(deleted bool) {
-		id, _ := strconv.Atoi(data[i.Row][0].(string))
+		if deleted {
+			id, _ := strconv.Atoi(data[i.Row][0].(string))
 
-		err := app.DB.DeleteHolding(int64(id))
-		if err != nil {
-			app.ErrorLog.Println(err)
+			err := app.DB.DeleteHolding(int64(id))
+			if err != nil {
+				app.ErrorLog.Println(err)
+			}
 		}
 
 		app.refreshHoldingsTable()
